all: add ISC license headers to all files and remove the LICENSE file
All the software I wrote here is distributed under the terms of the
OpenBSD variant of the ISC license. For easier tracking of copyright
of specific code, each file has its own copyright notice listing
specific copyright owners. This makes it easier both to copy code
out of this repository, and to import 3rd party code.